+++ Helix/src/Helix/Input/MouseCode.hpp
@@ -2,7 +2,7 @@
 
 namespace hlx::input
 {
-	enum MouseCode : int
+	enum MouseCode : uint16_t
 	{
 		Button0 = 0, 
 		Button1 = 1,
